v.13.0.0 - MyJD partially working
